refactor(dashboard): use client-side callback for sidebar link colours
